@@ -1,47 +1,68 @@
 using System;
 using System.IO;
-using System.Net.Http;
-using System.Net;
-using System.Threading.Tasks;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Host;
-using Microsoft.Azure.WebJobs.Extensions.Http;
+using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
-using static System.Net.Mime.MediaTypeNames;
 
-using Newtonsoft.Json;
+using Azure.Storage.Blobs;
+using Azure.Storage.Blobs.Models;
+using System.Security.Cryptography;
+using Azure.Storage.Blobs.Specialized;
 using System.Text;
-using System.Collections.Generic;
-using Microsoft.Azure.Storage.Blob;
 
 namespace Api.FileOps
 {
-
-    
     public class FileChanged
     {
-
         const string MD5Key = "ParsedMD5";
+        private readonly ILogger _logger;
 
-        [FunctionName("FileChanged")]
-        public async Task Run(
-            [BlobTrigger("bigbeercontainer/{name}", Connection = "BigBeerStorageAccount")] ICloudBlob blob,
+        public FileChanged(ILoggerFactory loggerFactory)
+        {
+            _logger = loggerFactory.CreateLogger<FileChanged>();
+        }
+
+        [Function("FileChanged")]
+        public async Task Run([BlobTrigger("bigbeercontainer/{name}", Connection = "BigBeerStorageAccount")]
+            ReadOnlyMemory<byte> blob,
             //[ServiceBus("process", Connection = "sbcss", EntityType = EntityType.Queue)] ICollector queueCollector,
             string name,
             Uri uri, // blob primary location
-            IDictionary<string, string> metaData, // user-defined blob metadata
-            BlobProperties properties, // blob system properties, e.g. LastModified
-            ILogger log)
+            BlobProperties properties,
+            IDictionary<string, string> metaData) // user-defined blob metadata
         {
-            if (metaData != null && metaData.ContainsKey(MD5Key) && metaData[MD5Key] == properties.ContentMD5)
+            byte[] md5 = MD5.HashData(blob.ToArray());
+            string sMD5 = Encoding.UTF8.GetString(md5);
+            BlobClient blobClient = new(uri);
+
+            string storageConnectionString = System.Environment.GetEnvironmentVariable("BigBeerStorageAccount") ?? String.Empty;
+            try
             {
-                // old content
-            } else
+                var blobServiceClient = new BlobServiceClient(storageConnectionString);
+                var cloudBlobContainer = blobServiceClient.GetBlobContainerClient("bigbeercontainer");
+                await cloudBlobContainer.CreateIfNotExistsAsync();
+
+                BlockBlobClient cloudBlockBlob = cloudBlobContainer.GetBlockBlobClient(name);
+
+                //_logger.LogInformation($"C# Blob trigger function Processed blob\n Name: {path}");
+                if (metaData != null && metaData.ContainsKey(MD5Key) && metaData[MD5Key] == sMD5)
+                {
+                    // old content
+                }
+                else
+                {
+
+                    var metadata = new Dictionary<string, string>
+                    {
+                        { MD5Key, sMD5 ?? String.Empty }
+                    };
+                    await cloudBlockBlob.SetMetadataAsync(metadata);
+                }
+            } catch (Exception e)
             {
-                blob.Metadata[MD5Key] = properties.ContentMD5;
-                await blob.SetMetadataAsync();
+                _logger.LogError(e, "Error checking upload");
             }
+
+               
         }
     }
 }
-
