@@ -1,52 +1,44 @@
-using System.IO;
-using System.Linq;
+using System.Collections.Generic;
 using System.Net;
-using System.Threading.Tasks;
 using BigBeerData.Shared;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
-using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Enums;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using Microsoft.OpenApi.Models;
-using Newtonsoft.Json;
 
 namespace Api.Locations
 {
-	public class CurrentLocation
-	{
-		private readonly BigBeerContext _context;
-		private readonly ILogger<LocationStyles> _logger;
+    public class CurrentLocation
+    {
 
-		public CurrentLocation(BigBeerContext context, ILogger<LocationStyles> log)
-		{
-			_context = context;
-			_logger = log;
-		}
+        private readonly BigBeerContext _context;
+        private readonly ILogger _logger;
 
-		[FunctionName("CurrentLocation")]
-		[OpenApiOperation(operationId: "Run")]
-		[OpenApiSecurity("function_key", SecuritySchemeType.ApiKey, Name = "code", In = OpenApiSecurityLocationType.Query)]
-		[OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "application/json", bodyType: typeof(string), Description = "The OK response")]
-		public async Task<IActionResult> Run(
-						[HttpTrigger(AuthorizationLevel.Function, "get", Route = "location/current")]
-			HttpRequest req)
-		{
-			_logger.LogDebug("Get Current Location called");
-			var establishments = _context.Locations.Where(l => l.Establishments.Any());
+        public CurrentLocation(BigBeerContext context, ILoggerFactory loggerFactory)
+        {
+            _logger = loggerFactory.CreateLogger<LocationStyle>();
+            _context = context;
+        }
 
-			var result = establishments.Select(s =>
-				new AreaLocation
-				{
-					ID = s.LocationId,
-					Name = s.LocationName
-				});
+        [Function("CurrentLocation")]
+        public async Task<HttpResponseData> Run([HttpTrigger(AuthorizationLevel.Function, "get", Route = "location/current")] 
+        HttpRequestData req)
+        {
+            var establishments = _context.Locations.Where(l => l.Establishments.Any());
 
-			return new OkObjectResult(await result.ToListAsync());
-		}
-	}
+            var result = await establishments.Select(s =>
+                new AreaLocation
+                {
+                    ID = s.LocationId,
+                    Name = s.LocationName
+                }).ToListAsync();
+
+            _logger.LogInformation("Get Current Location Called.");
+
+            var response = req.CreateResponse(HttpStatusCode.OK);
+            await response.WriteAsJsonAsync(result);
+
+            return response;
+        }
+    }
 }
-
