@@ -1,40 +1,35 @@
-﻿using Microsoft.Azure.Functions.Extensions.DependencyInjection;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
-using System;
-using System.Net.Http.Headers;
-
-using Microsoft.Azure.Storage.Blob;
-
-using Microsoft.Azure.Storage;
 
 using BigBeerData.Shared;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 
-[assembly: FunctionsStartup(typeof(Api.Startup))]
-namespace Api
-{
-	public class Startup : FunctionsStartup
-	{
-		public override void Configure(IFunctionsHostBuilder builder)
-		{
-			builder.Services.AddDbContext<BigBeerContext>(opt =>
-									opt.UseSqlServer(
-											builder.GetContext().Configuration["DBConnection"]
-									));
-			
+using System.Net.Http.Headers;
 
-			builder.Services.AddHttpClient("BeerBot", config =>
-								{
-									var productValue = new ProductInfoHeaderValue("BBD", "1.0");
-									var commentValue = new ProductInfoHeaderValue("(Hi)");
+IConfiguration config = new ConfigurationBuilder()
+    .AddEnvironmentVariables()
+    .Build();
 
-									config.BaseAddress = new Uri("https://api.untappd.com/v4/");
+var host = new HostBuilder()
+		.ConfigureFunctionsWorkerDefaults()
+        .ConfigureServices(builder => {
+            builder.AddDbContext<BigBeerContext>(opt =>
+                                    opt.UseSqlServer(
+                                        config["DBConnection"]
+                                    ));
+            builder.AddHttpClient("BeerBot", config =>
+             {
+                 var productValue = new ProductInfoHeaderValue("BBD", "1.0");
+                 var commentValue = new ProductInfoHeaderValue("(Hi)");
 
-									config.DefaultRequestHeaders.UserAgent.Add(productValue);
-									config.DefaultRequestHeaders.UserAgent.Add(commentValue);
-								});
+                 config.BaseAddress = new Uri("https://api.untappd.com/v4/");
 
-		}
-	}
-}
+                 config.DefaultRequestHeaders.UserAgent.Add(productValue);
+                 config.DefaultRequestHeaders.UserAgent.Add(commentValue);
+             });
+        })
+		.Build();
+
+host.Run();
