@@ -2,8 +2,8 @@
 {
     public class BlobDto
     {
-        public string Name { get; set; }
-        public string FileUrl { get; set; }
-        public string ContentType { get; set; }
+        public string Name { get; set; } = default!;
+        public string FileUrl { get; set; } = default!;
+        public string ContentType { get; set; } = default!;
     }
 }
