@@ -2,8 +2,8 @@
 {
     public class ContentDto
     {
-        public Stream Content { get; set; }
-        public string Name { get; set; }
-        public string ContentType { get; set; }
+        public Stream Content { get; set; } = default!;
+        public string Name { get; set; } = default!;
+        public string ContentType { get; set; } = default!;
     }
 }
