@@ -52,8 +52,8 @@
 
 
   public bool loaded = false;
-  public Client.Pages.Index mainPage;
-  public List<AreaLocation> locations;
+  public Client.Pages.Index mainPage = default!;
+  public List<AreaLocation> locations = default!;
   public int selected = 1;
   public bool updating = false;
 
