@@ -32,7 +32,7 @@
 @code {
 
 	[Parameter]
-	public MainLayout parent { get; set; }
+	public MainLayout parent { get; set; } = default!;
 
 	private string? selected(int currentId)
 	{
